Prepare for online e2e tests
Currently, none of our e2e tests require network access. That is good
when running them on the build server, as tests will not randomly fail
due to temporary network issues. However, during development it can be
useful to have a small set of automated e2e tests that perform network
operations. F.ex. we regularly want to test that app installation works.

In preparation of online e2e tests, I am moving our current e2e tests
below the test/offline directory. Then we can have online tests in
test/online, which allows us to easily specify to jest that we want to
run either offline or online tests (or both).
