+++ test/offline/fixtures/one-local-app-to-be-extracted/.nrfconnect-apps/apps.json
