+++ test/offline/fixtures/one-local-app-to-be-extracted/pc-nrfconnect-foo-1.2.3.tgz
