+++ test/offline/fixtures/one-local-app-unsupported-engine/.nrfconnect-apps/apps.json
