+++ test/offline/fixtures/one-local-app-unsupported-engine/.nrfconnect-apps/local/pc-nrfconnect-test/index.js
