+++ test/offline/fixtures/one-local-app-unsupported-engine/.nrfconnect-apps/package.json
