+++ test/offline/fixtures/one-local-app-unsupported-engine/.nrfconnect-apps/updates.json
