+++ test/offline/fixtures/one-local-app-without-engine/.nrfconnect-apps/local/pc-nrfconnect-test/index.js
