+++ test/offline/fixtures/one-local-app-without-engine/.nrfconnect-apps/package.json
