+++ test/offline/fixtures/one-local-app/.nrfconnect-apps/apps.json
