+++ test/offline/fixtures/one-local-app/.nrfconnect-apps/local/pc-nrfconnect-test/index.js
