+++ test/offline/fixtures/one-local-app/.nrfconnect-apps/package.json
