+++ test/offline/fixtures/one-local-app/.nrfconnect-apps/updates.json
