+++ test/offline/fixtures/one-official-app-installed/.nrfconnect-apps/updates.json
