+++ test/offline/fixtures/one-official-app-not-installed/.nrfconnect-apps/local/.gitignore
