+++ test/offline/fixtures/one-official-app-not-installed/.nrfconnect-apps/updates.json
