+++ test/offline/fixtures/one-official-app-upgradable/.nrfconnect-apps/apps.json
