+++ test/offline/fixtures/one-official-app-upgradable/.nrfconnect-apps/package.json
