+++ test/setup/index.js
@@ -50,7 +50,6 @@ if (process.platform === 'win32') {
 function startElectronApp(extraArgs) {
     const args = [
         projectPath,
-        '--skip-update-apps',
         '--skip-update-core',
         '--skip-splash-screen',
     ];
